Modified: Create useScrollAnchor hook to handle scrolling functionality
